requirements.txt
azure-cognitiveservices-speech==1.18.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
filelock==3.0.12
huggingface-hub==0.0.16
idna==3.2
joblib==1.0.1
nltk==3.6.2
numpy==1.21.2
packaging==21.0
Pillow==8.3.2
pyparsing==2.4.7
pytesseract==0.3.8
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
transformers==4.10.0
typing-extensions==3.10.0.2
urllib3==1.26.6
Wand==0.6.7
